Expose ArgoCD's `ignoreDifferences` as a component parameter
We allow users to configure ArgoCD's diffing customization through the
hierarchy in parameter `argocd_ignore_differences`. This is helpful when
deploying some resource which is mutated by an admission webhook through
the adhoc-configurations component.

The new parameter is an object, to allow users to modify or remove
existing configurations in the hierarchy. However, the component
completely ignores the object keys, and uses only the non-null values to
create the list for `spec.ignoreDifferences` in the ArgoCD application
resource.

Closes #25